refactor: AttributeKey sku -> constant 패키지로 이동
