+++ src/main/java/com/fittura/domain/product/product/dto/request/AttributeCreateReqDto.java
@@ -1,15 +1,15 @@
 package com.fittura.domain.product.product.dto.request;
 
-import com.fittura.domain.product.sku.constant.AttributeKey;
+import com.fittura.domain.product.product.constant.AttributeKey;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-@Schema(description = "SKU 요소 생성 요청 DTO")
+@Schema(description = "Product 요소 생성 요청 DTO")
 public record AttributeCreateReqDto(
 
-    @Schema(example = "SIZE")
+    @Schema(example = "SIZE_LABEL")
     @NotNull
     AttributeKey attributeKey,
