+++ src/main/java/com/fittura/domain/product/sku/dto/responseDto/ProductAttributeResDto.java
@@ -1,6 +1,6 @@
 package com.fittura.domain.product.sku.dto.responseDto;
 
-import com.fittura.domain.product.sku.constant.AttributeKey;
+import com.fittura.domain.product.product.constant.AttributeKey;
 
 public record ProductAttributeResDto(
     Long id,
